add parsing for entire programs
